Refactor, fix logging.
