+++ txcache/diagnosis.go
@@ -1,9 +1,25 @@
 package txcache
 
 import (
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
+	"strings"
+
 	"github.com/multiversx/mx-chain-core-go/core"
 	logger "github.com/multiversx/mx-chain-logger-go"
 )
+
+type printedTransaction struct {
+	Hash       string  `json:"hash"`
+	Nonce      uint64  `json:"nonce"`
+	PPU        float64 `json:"ppu"`
+	GasPrice   uint64  `json:"gasPrice"`
+	GasLimit   uint64  `json:"gasLimit"`
+	Sender     string  `json:"sender"`
+	Receiver   string  `json:"receiver"`
+	DataLength int     `json:"dataLength"`
+}
 
 // Diagnose checks the state of the cache for inconsistencies and displays a summary, senders and transactions.
 func (cache *TxCache) Diagnose(_ bool) {
@@ -48,14 +64,47 @@ func (cache *TxCache) diagnoseTransactions() {
 	}
 
 	transactions := cache.getAllTransactions()
-
 	if len(transactions) == 0 {
 		return
 	}
 
 	numToDisplay := core.MinInt(diagnosisMaxTransactionsToDisplay, len(transactions))
 	logDiagnoseTransactions.Trace("diagnoseTransactions", "numTransactions", len(transactions), "numToDisplay", numToDisplay)
-	logDiagnoseTransactions.Trace(marshalTransactionsToNewlineDelimitedJson(transactions[:numToDisplay]))
+	logDiagnoseTransactions.Trace(marshalTransactionsToNewlineDelimitedJson(transactions[:numToDisplay], "diagnoseTransactions"))
+}
+
+// marshalTransactionsToNewlineDelimitedJson converts a list of transactions to a newline-delimited JSON string.
+// Note: each line is indexed, to improve readability. The index is easily removable for if separate analysis is needed.
+func marshalTransactionsToNewlineDelimitedJson(transactions []*WrappedTransaction, linePrefix string) string {
+	builder := strings.Builder{}
+	builder.WriteString("\n")
+
+	for i, wrappedTx := range transactions {
+		printedTx := convertWrappedTransactionToPrintedTransaction(wrappedTx)
+		printedTxJson, _ := json.Marshal(printedTx)
+
+		builder.WriteString(fmt.Sprintf("%s#%d: ", linePrefix, i))
+		builder.WriteString(string(printedTxJson))
+		builder.WriteString("\n")
+	}
+
+	builder.WriteString("\n")
+	return builder.String()
+}
+
+func convertWrappedTransactionToPrintedTransaction(wrappedTx *WrappedTransaction) *printedTransaction {
+	transaction := wrappedTx.Tx
+
+	return &printedTransaction{
+		Hash:       hex.EncodeToString(wrappedTx.TxHash),
+		Nonce:      transaction.GetNonce(),
+		Receiver:   hex.EncodeToString(transaction.GetRcvAddr()),
+		Sender:     hex.EncodeToString(transaction.GetSndAddr()),
+		GasPrice:   transaction.GetGasPrice(),
+		GasLimit:   transaction.GetGasLimit(),
+		DataLength: len(transaction.GetData()),
+		PPU:        wrappedTx.PricePerUnit,
+	}
 }
 
 func (cache *TxCache) diagnoseSelection() {
@@ -63,22 +112,18 @@ func (cache *TxCache) diagnoseSelection() {
 		return
 	}
 
-	transactions := cache.doSelectTransactions(
-		logDiagnoseSelection,
-		diagnosisSelectionGasRequested,
-	)
-
-	displaySelectionOutcome(logDiagnoseSelection, transactions)
+	transactions := cache.doSelectTransactions(diagnosisSelectionGasRequested)
+	displaySelectionOutcome(logDiagnoseSelection, "diagnoseSelection", transactions)
 }
 
-func displaySelectionOutcome(contextualLogger logger.Logger, selection []*WrappedTransaction) {
+func displaySelectionOutcome(contextualLogger logger.Logger, linePrefix string, transactions []*WrappedTransaction) {
 	if contextualLogger.GetLevel() > logger.LogTrace {
 		return
 	}
 
-	if len(selection) > 0 {
+	if len(transactions) > 0 {
 		contextualLogger.Trace("displaySelectionOutcome - transactions (as newline-separated JSON):")
-		contextualLogger.Trace(marshalTransactionsToNewlineDelimitedJson(selection))
+		contextualLogger.Trace(marshalTransactionsToNewlineDelimitedJson(transactions, linePrefix))
 	} else {
 		contextualLogger.Trace("displaySelectionOutcome - transactions: none")
 	}
